+++ Runtime/ContinuousDistributions/BatesDistribution/Generators/BatesGenerator.cs
@@ -13,10 +13,12 @@ namespace Zor.RandomGenerators.ContinuousDistributions
 	[Serializable]
 	public sealed class BatesGenerator : IBatesGenerator
 	{
+#pragma warning disable CS0649
 		[SerializeField] private float m_Mean = BatesDistribution.DefaultMean;
 		[SerializeField] private float m_Deviation = BatesDistribution.DefaultDeviation;
 		[SerializeField, Tooltip("How many independent and identically distributed random variables are generated.")]
 		private byte m_Iids = BatesDistribution.DefaultIids;
+#pragma warning restore CS0649
 
 		/// <summary>
 		/// Creates a <see cref="BatesGenerator"/> with the default parameters.